atomic: change atomic_t to long
Change the atomic_t type from int to long. For all chips used for EC,
this change shouldn't impact the generated code, except for the string
needed fot the print function.

The change caused by a change in Zephyr upstream (atomic_t from int to
long).

BUG=b:207082842
Cq-Depend: chromium:3330355
TEST=make buildall && zmake testall
BRANCH=main

+++ include/atomic_t.h
@@ -9,7 +9,7 @@
 #define __CROS_EC_ATOMIC_T_H
 
 #ifndef CONFIG_ZEPHYR
-typedef int atomic_t;
+typedef long atomic_t;
 typedef atomic_t atomic_val_t;
 #else
 #include <sys/atomic.h>
